server-side update for the "Add Task" feature

+++ server/src/app.js
@@ -31,10 +31,12 @@ import { deleteUserIfExpired } from './models/User.js';
 app.use(cookieParser());
 
 // We use cron to schedule tasks
-cron.schedule('00 12 * * *', () => {
-  // running a task every day at 12:00
-  deleteUserIfExpired();
-});
+if (process.env.NODE_ENV !== 'test') {
+  cron.schedule('00 12 * * *', () => {
+    // running a task every day at 12:00
+    deleteUserIfExpired();
+  });
+}
 
 /****** Attach routes ******/
 /**
